requirements.txt
# Python dependencies for Resume Relevance Analysis System

# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# File handling
python-multipart==0.0.6
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3

# NLP libraries
spacy==3.7.2
nltk==3.8.1
sentence-transformers==2.2.2
transformers==4.35.2

# Machine Learning
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4

# Text processing
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
textdistance==4.6.0

# LLM and vector stores
langchain==0.0.340
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.3.7
chromadb==0.4.18

# HTTP client
httpx==0.25.2
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0

# Visualization (for Streamlit dashboard)
plotly==5.17.0
matplotlib==3.8.2

# Async support
asyncio==3.4.3
aiofiles==23.2.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

# Additional utilities
tqdm==4.66.1
python-dateutil==2.8.2